Move the NumericInput component to 'Share' folder
